chore: refactor subgraph codes
